requirements.txt
blis==0.7.11
catalogue==1.0.2
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
charset-normalizer==3.4.1
ChatterBot==1.0.8
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
cymem==2.0.11
dnspython==2.3.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
Flask==2.2.5
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.0.5
greenlet==3.1.1
idna==3.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.5
langcodes==3.3.0
MarkupSafe==2.1.5
mathparse==0.1.2
murmurhash==1.0.12
numpy==1.21.6
packaging==24.0
pathy==0.10.3
plac==1.1.3
preshed==3.0.9
psycopg2==2.9.9
pycparser==2.21
pydantic==1.8.2
PyJWT==2.8.0
pymongo==4.7.3
python-dateutil==2.8.2
pytz==2025.1
requests==2.31.0
six==1.17.0
smart-open==6.4.0
spacy==2.3.9
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.3.0
srsly==1.0.7
thinc==7.4.6
tqdm==4.67.1
typer==0.4.2
typing_extensions==4.7.1
urllib3==2.0.7
wasabi==0.10.1
Werkzeug==2.2.3
wincertstore==0.2
zipp==3.15.0
